#11 fix phpmd failed
Just add a abstract class for project model, the abstract model
contains all relations of project model.
